+++ src/blockchain/logic.js
@@ -100,7 +100,18 @@ const generateNextBlock = (blockData, blockchain, Block) => {
   const nextHash = helpers.crypto.calculateHash(nextIndex, previousBlock.hash, nextTimestamp, blockData);
   return new Block(nextIndex, previousBlock.hash, nextTimestamp, blockData, nextHash);
 };
-export { MessageType }
+
+const getGenesisEMRBlock = () => {
+  return {
+    "index": 0,
+    "previousHash": `0`,
+    "timestamp": 1465154705,
+    "hash": `816534932c2b7154836da6afc367695e6337db8a921823784c14378abed4f7d7`,
+    "data": `U2FsdGVkX19rnP7i0ysECsV/l7QeKI8oN7oXOyM4vN4=`
+  };
+};
+
+export { MessageType, getGenesisEMRBlock }
 export default {
   MessageType,
   getLatestBlock,
